Split Simplify.scala in 15 files
