+++ compiler/src/dotty/tools/dotc/transform/localopt/Optimisation.scala
@@ -1,0 +1,20 @@
+package dotty.tools.dotc
+package transform.localopt
+
+import core.Contexts.Context
+import ast.tpd.Tree
+
+trait Optimisation {
+
+  /** Run first to gather information on Trees (using mutation) */
+  def visitor: Tree => Unit
+
+  /** Does the actual Tree => Tree transformation, possibly using a different
+   *  context from the one using in Optimisation.
+   */
+  def transformer(localCtx: Context): Tree => Tree
+
+  def name: String = this.getClass.getName.init
+
+  val NoVisitor: Tree => Unit = _ => ()
+}
